type and schema one export per file
