+++ packages/twenty-server/src/engine/metadata-modules/ai/ai-models/types/ai-provider-auth-type.schema.ts
@@ -1,0 +1,3 @@
+import { z } from 'zod';
+
+export const aiProviderAuthTypeSchema = z.enum(['key', 'credentials', 'role']);
