+++ packages/twenty-server/src/engine/metadata-modules/ai/ai-models/types/ai-provider-config.type.ts
@@ -1,27 +1,7 @@
 import { z } from 'zod';
 
-import { AI_SDK_PACKAGES, DATA_RESIDENCY_KEYS } from 'twenty-shared/ai';
-
-import { aiProviderAuthTypeSchema } from 'src/engine/metadata-modules/ai/ai-models/types/ai-provider-auth-type.type';
-import {
-  aiProviderModelConfigSchema,
-  type AiProviderModelConfig,
-} from 'src/engine/metadata-modules/ai/ai-models/types/ai-provider-model-config.type';
-
-export const aiProviderConfigSchema = z.object({
-  npm: z.enum(AI_SDK_PACKAGES),
-  name: z.string().optional(),
-  label: z.string().optional(),
-  authType: aiProviderAuthTypeSchema.optional(),
-  apiKey: z.string().optional(),
-  baseUrl: z.string().optional(),
-  region: z.string().optional(),
-  dataResidency: z.enum(DATA_RESIDENCY_KEYS).optional(),
-  accessKeyId: z.string().optional(),
-  secretAccessKey: z.string().optional(),
-  sessionToken: z.string().optional(),
-  models: z.array(aiProviderModelConfigSchema).optional(),
-});
+import { aiProviderConfigSchema } from 'src/engine/metadata-modules/ai/ai-models/types/ai-provider-config.schema';
+import { type AiProviderModelConfig } from 'src/engine/metadata-modules/ai/ai-models/types/ai-provider-model-config.type';
 
 export type AiProviderConfig = Omit<
   z.infer<typeof aiProviderConfigSchema>,
