+++ packages/twenty-server/src/engine/metadata-modules/ai/ai-models/types/ai-providers-config.type.ts
@@ -1,13 +1,3 @@
-import { z } from 'zod';
-
-import {
-  aiProviderConfigSchema,
-  type AiProviderConfig,
-} from 'src/engine/metadata-modules/ai/ai-models/types/ai-provider-config.type';
-
-export const aiProvidersConfigSchema = z.record(
-  z.string(),
-  aiProviderConfigSchema,
-);
+import { type AiProviderConfig } from 'src/engine/metadata-modules/ai/ai-models/types/ai-provider-config.type';
 
 export type AiProvidersConfig = Record<string, AiProviderConfig>;
