+++ packages/twenty-server/src/engine/metadata-modules/ai/ai-models/types/long-context-cost.schema.ts
@@ -1,0 +1,9 @@
+import { z } from 'zod';
+
+export const longContextCostSchema = z.object({
+  inputCostPerMillionTokens: z.number(),
+  outputCostPerMillionTokens: z.number(),
+  cachedInputCostPerMillionTokens: z.number().optional(),
+  cacheCreationCostPerMillionTokens: z.number().optional(),
+  thresholdTokens: z.number(),
+});
